pyproject.toml
[tool.poetry]
name = "healthfinance-fastapi"
version = "0.1.0"
description = "Private rest api to list financial customer data."
authors = ["Abraão de Andrade Farias Reis"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pylint = "^3.0.3"
pre-commit = "^3.6.0"
sqlalchemy = "^2.0.23"
pytest = "^7.4.3"
bcrypt = "^4.1.2"
flask = "^3.0.0"
cerberus = "^1.3.5"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
